requirements.txt
colorama==0.4.6
filelock==3.16.1
fsspec==2025.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
pillow==10.4.0
python-mnist==0.7
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.13.3
threadpoolctl==3.5.0
torch==1.8.0
torchvision==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
